refactor: cutout OptionBase as single file and replace it
https://github.com/akky-xxxx/eslint-plugin-strict-check/issues/285
