+++ src/shared/types/OptionBase/index.ts
@@ -1,0 +1,3 @@
+export type OptionBase<O extends Record<string, unknown>> = {
+  [K in keyof O]?: O[K] | unknown
+}
